duplicate the PlanTest with catalyst/plans/PlanTest
